backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==23.5
scipy>=1.13.0
numpy>=1.26.4
python-dateutil==2.9.0
